+++ src/viewPages/SelectionSpotsWidget.cpp
@@ -30,7 +30,7 @@ SelectionSpotsWidget::SelectionSpotsWidget(const UserSelection::STDataFrame &dat
     // populate
     for (uword i = 0; i < data.counts.n_rows; ++i) {
         const auto spot_str = data.spots.at(i);
-        const float count = sum(data.counts.col(i));
+        const double count = sum(data.counts.col(i));
         const QString count_str = QString::number(count);
         QStandardItem *spot_item = new QStandardItem(spot_str);
         spot_item->setData(spot_str, Qt::UserRole);
